Error check the wiki generation before posting.
In the regenerate-wiki.sh, check to make sure that the
top level wiki document generation script doesn't fail.
If it fails for any reason, we don't publish.  Also check
to make sure that the assignments page (see page ids in the
conf files) also generate without errors.

Additionally check that there are no python tracebacks in
the output.

One more improvement being worked on is to store the generated
wiki markdown file and compare against the previous version.
If no changes are detected don't push the markdown.

related-to https://github.com/redhat-performance/quads/issues/52

+++ bin/regenerate-wiki.sh
@@ -35,8 +35,20 @@ fi
 tmpfile=$(mktemp /tmp/wikimarkdownXXXXX)
 
 $bindir/create-input.sh 1>$tmpfile  2>&1
+if [ $? -gt 0 ]; then
+    exit 1
+fi
+if grep -q Traceback $tmpfile ; then
+    exit 1
+fi
 $bindir/racks-wiki.py --markdown $tmpfile --wp-url http://$wp_wiki/xmlrpc.php --wp-username  $wp_username --wp-password  $wp_password --page-title "$wp_wiki_main_title" --page-id $wp_wiki_main_page_id
 $bindir/create-input-assignments.sh 1>$tmpfile  2>&1
+if [ $? -gt 0 ]; then
+    exit 1
+fi
+if grep -q Traceback $tmpfile ; then
+    exit 1
+fi
 $bindir/racks-wiki.py --markdown $tmpfile --wp-url http://$wp_wiki/xmlrpc.php --wp-username  $wp_username --wp-password  $wp_password --page-title "$wp_wiki_assignments_title" --page-id $wp_wiki_assignments_page_id
 rm -f $tmpfile $lockfile
